adreno_tz: Adjust GPU target frequency calculation for overclocked panel Based on an idea by Sultan Alsawaf, however we arent using dynamic refresh rate implementation.  Since we are shipping the kernel at 75HZ, let's multiply the value by 1.25 instead.
original commit message:

It seems that the black-box GPU frequency selection algorithm is unaware
of high refresh rates, and assumes a refresh rate of 60 Hz for its
target frequencies. This results in frequencies that are too low being
selected when a higher refresh rate, such as 90 Hz, is used.

To remedy this, multiply the GPU's busy time by 1.5 when 90 Hz is used,
since the graphical load of 90 Hz should be about 1.5x greater than the
load of 60 Hz. This makes msm-adreno-tz select the right frequencies
needed to avoid stuttering.

Signed-off-by: dlwlrma123 <alexfinhart@gmail.com>